Fragment src::channel_handler::types into many files
