+++ src/channel_handler/types/accept_transaction_state.rs
@@ -1,0 +1,8 @@
+use crate::referee::RefereeOnChainTransaction;
+
+#[derive(Debug)]
+pub enum AcceptTransactionState {
+    Determined(Box<RefereeOnChainTransaction>),
+    Waiting,
+    Finished,
+}
