+++ src/channel_handler/types/coin_data_for_reward.rs
@@ -1,0 +1,8 @@
+use crate::common::types::CoinString;
+
+pub struct CoinDataForReward {
+    pub coin_string: CoinString,
+    // parent: CoinID,
+    // puzzle_hash: PuzzleHash,
+    // amount: Amount,
+}
