+++ src/channel_handler/types/coin_identification_by_puzzle_hash.rs
@@ -1,0 +1,7 @@
+use crate::common::types::{Amount, PuzzleHash};
+
+#[derive(Debug)]
+pub enum CoinIdentificationByPuzzleHash {
+    Reward(PuzzleHash, Amount),
+    Game(PuzzleHash, Amount),
+}
