+++ src/channel_handler/types/on_chain_game_coin.rs
@@ -1,0 +1,7 @@
+use crate::common::types::{CoinString, GameID};
+
+#[derive(Debug, Clone)]
+pub struct OnChainGameCoin {
+    pub game_id_up: GameID,
+    pub coin_string_up: Option<CoinString>,
+}
