+++ src/channel_handler/types/on_chain_game_state.rs
@@ -1,0 +1,10 @@
+use super::accept_transaction_state::AcceptTransactionState;
+use crate::common::types::{GameID, PuzzleHash};
+
+#[derive(Debug)]
+pub struct OnChainGameState {
+    pub game_id: GameID,
+    pub puzzle_hash: PuzzleHash,
+    pub our_turn: bool,
+    pub accept: AcceptTransactionState,
+}
